Avoid using deprecated apis
Use new DialogFragment and PreferenceFragment. Also organise the code in different packages.
