+++ app/src/main/java/org/laurentsebag/wifitimer/activities/MainActivity.java
@@ -16,22 +16,20 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.laurentsebag.wifitimer;
+package org.laurentsebag.wifitimer.activities;
 
-import android.app.Activity;
-import android.app.AlertDialog;
-import android.app.Dialog;
 import android.content.Intent;
-import android.content.res.Resources;
 import android.os.Bundle;
-import android.text.method.LinkMovementMethod;
-import android.util.TypedValue;
+import android.support.v4.app.FragmentActivity;
 import android.view.View;
-import android.widget.TextView;
 
-public class InfoActivity extends Activity implements View.OnClickListener {
+import org.laurentsebag.wifitimer.fragments.AboutDialogFragment;
+import org.laurentsebag.wifitimer.R;
 
-    private static final int ABOUT_DIALOG = 0;
+public class MainActivity extends FragmentActivity implements View.OnClickListener {
+
+    private static final String ABOUT_DIALOG = "about_dialog";
+    private AboutDialogFragment aboutDialog;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -40,43 +38,21 @@ public class InfoActivity extends Activity implements View.OnClickListener {
         findViewById(R.id.done).setOnClickListener(this);
         findViewById(R.id.settings).setOnClickListener(this);
         findViewById(R.id.about).setOnClickListener(this);
+        aboutDialog = (new AboutDialogFragment());
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public void onClick(View v) {
         switch (v.getId()) {
             case R.id.done:
                 finish();
                 break;
             case R.id.about:
-                showDialog(ABOUT_DIALOG);
+                aboutDialog.show(getSupportFragmentManager(), ABOUT_DIALOG);
                 break;
             case R.id.settings:
                 Intent i = new Intent(this, SettingsActivity.class);
                 startActivity(i);
                 break;
-        }
-    }
-
-    @Override
-    protected Dialog onCreateDialog(int id) {
-        switch (id) {
-            case ABOUT_DIALOG:
-                Resources r = getResources();
-                int padding = (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, 15, r.getDisplayMetrics());
-                AlertDialog.Builder builder = new AlertDialog.Builder(this);
-                builder.setTitle(R.string.about_dialog_title);
-                TextView content = new TextView(this);
-                content.setText(R.string.about_dialog_content);
-                content.setTextSize(TypedValue.COMPLEX_UNIT_DIP, 16);
-                content.setMovementMethod(LinkMovementMethod.getInstance());
-                content.setPadding(padding, padding, padding, padding);
-                builder.setView(content);
-                return builder.create();
-            default:
-                return super.onCreateDialog(id);
         }
     }
 }
