+++ app/src/main/java/org/laurentsebag/wifitimer/activities/SettingsActivity.java
@@ -1,6 +1,6 @@
 /*-
- *  Copyright (C) 2011 Laurent Sebag   
- *  
+ *  Copyright (C) 2011 Laurent Sebag
+ *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
  *  the Free Software Foundation, either version 3 of the License, or
@@ -15,19 +15,21 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.laurentsebag.wifitimer;
+package org.laurentsebag.wifitimer.activities;
 
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.Intent;
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v4.app.FragmentActivity;
 
-/**
- * Receives alarm broadcast to turn the wifi back on.
- */
-public class AlarmReceiver extends BroadcastReceiver {
+import org.laurentsebag.wifitimer.fragments.SettingsFragment;
 
+public class SettingsActivity extends FragmentActivity {
     @Override
-    public void onReceive(Context context, Intent intent) {
-        RadioUtils.setWifiStateBack(context);
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        getSupportFragmentManager().beginTransaction()
+                .replace(android.R.id.content, new SettingsFragment())
+                .commit();
     }
 }
