+++ app/src/main/res/xml/settings.xml
@@ -3,13 +3,13 @@
     android:title="@string/settings_activity_label">
 
     <ListPreference
-        android:key="timer_usage"
-        android:title="@string/preference_timer_usage_title"
-        android:summary="@string/preference_timer_usage_summary"
+        android:defaultValue="on_wifi_deactivation"
+        android:dialogTitle="@string/preference_timer_usage_dialog_title"
         android:entries="@array/preference_timer_usage_entries"
         android:entryValues="@array/preference_timer_usage_entryvalues"
-        android:dialogTitle="@string/preference_timer_usage_dialog_title"
-        android:defaultValue="on_wifi_deactivation"
-        android:persistent="true"/>
+        android:key="timer_usage"
+        android:persistent="true"
+        android:summary="@string/preference_timer_usage_summary"
+        android:title="@string/preference_timer_usage_title"/>
 
 </PreferenceScreen>
